requirements.txt
opencv-python==4.5.5.62
pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
mtcnn-opencv==1.0.2
fastapi==0.71.0
uvicorn==0.16.0
python-multipart==0.0.5
timm==0.4.5